==> R/00_data-cleaning.Rmd <==
---
title: "Database importation and modifications"
author : "Nina Prasil & Jeremy Wicquart"
date: "`r Sys.Date()`"
output: 
  html_document:
    theme: "cosmo"
    highlight: tango
    toc: true
    toc_float:
      collapsed: false
      smooth_scroll: false
    toc_depth: 4
---

```{r setup, include=FALSE}

knitr::opts_chunk$set(echo = TRUE, fig.align = "center")

```

# Defining functions and packages

```{r base}

# 1. Required packages ----

library(tidyverse) # Core tidyverse packages
library(sf) # To plot maps
sf_use_s2(FALSE) # Switch from S2 to GEOS

```

# 1. Reef area by country

```{r}

# 1. Load data ----

load(file = "./../data/reefs.eez.RData")

# 2. Calculate the reef area by country ----

data_reefs <- reefs.eez %>%
  # Re-assign TERRITORY1 depending on geopolitical issues
  mutate(SOVEREIGN1 = case_when(TERRITORY1 == "Glorioso Islands" ~ "France",
                                TERRITORY1 == "Chagos Archipelago" ~ "United Kingdom",
                                TRUE ~ SOVEREIGN1)) %>% 
  # Convert to sf
  st_as_sf() %>% 
  # Calculate the reef area and transform it from m2 to km2
  mutate(reef_area = as.numeric(st_area(.))/1000000) %>%
  # Make the sum of area by country
  st_drop_geometry() %>% 
  group_by(SOVEREIGN1) %>% 
  summarise(total_area_abs = sum(reef_area)) %>% 
  ungroup() %>% 
  drop_na(SOVEREIGN1) %>% # By doing so I remove a small portion of unassigned reef surface
  # Sort rows from higher reef area to lower
  arrange(-total_area_abs) %>% 
  # Calculate the relative contribution
  mutate(total_area_rel = total_area_abs*100/sum(total_area_abs),
         pos = row_number()) %>% 
  # Misc. modifications for formattable
  select(pos, SOVEREIGN1, total_area_abs, total_area_rel) %>% 
  mutate(total_area_abs = round(as.numeric(total_area_abs), 3),
         total_area_rel = round(as.numeric(total_area_rel), 3))

# 3. Save the data ----

write_csv2(data_reefs, "../figs/04_table-reef-surface-by-country.csv")

# 4. Remove useless objects ----

rm(reefs.eez)

```

# 2. Economic Exclusive Zone

```{r}

# 1. Load EEZ data ----

data_eez <- read_sf("../data/01_maps/World_EEZ_v11_20191118/eez_v11.shp")

# 2. Change SOVEREIGN1 for Gloriosos islands from Madagascar to France ----

data_eez <- data_eez %>% 
  mutate(SOVEREIGN1 = case_when(TERRITORY1 == "Glorioso Islands" ~ "France",
                                TERRITORY1 == "Chagos Archipelago" ~ "United Kingdom",
                                TRUE ~ SOVEREIGN1)) %>% 
  st_transform(crs = 4326)

```

# 3. Defining top 30

```{r}

# 1. Extract the top 30 countries in reef surface ----

top_30 <- data_reefs %>% 
  select(pos, SOVEREIGN1) %>% 
  slice(1:30)

# 2. Extract benthic GCRMN data corresponding to the top 30 ---- 

data_benthos <- read.csv2("../data/03-merge_all_all_all_benthos_NA.csv") %>% 
  # Remove some datasets (those leading to bias)
  filter(!(DatasetID %in% c("XLCA1", "XLCA2", "XLCA3", "XLCA4", "XLCA5"))) %>% 
  # Remove NA
  drop_na(Latitude, Longitude, Year) %>%
  # Set as sf object
  st_as_sf(., coords = c("Longitude", "Latitude"), crs = 4326) %>% 
  # Join with EEZ
  st_join(., data_eez) %>% 
  # Extract top 30 countries in reef surface
  filter(SOVEREIGN1 %in% unique(top_30$SOVEREIGN1))

# 3. Save the data ----

save(data_benthos, file = "./../data/00_filtered/00-data_benthos-top_30.RData")

```

# 4. Defining time range (summarized dataset)

```{r}

# 1. Load data ----

data_benthos_range <- data_benthos %>% 
  mutate(Longitude = st_coordinates(.)[,"X"],
         Latitude = st_coordinates(.)[,"Y"]) %>% 
  st_drop_geometry() %>% 
  group_by(Area, Country, Location, Site, Latitude, Longitude, SOVEREIGN1) %>% 
  summarise(interval_years = max(Year, na.rm = TRUE) - min(Year, na.rm = TRUE)) %>%
  ungroup() %>% 
  mutate(interval_class = cut(interval_years, 
                              breaks = c(-Inf, 1, 5, 10, 15, Inf),
                              labels = c("1", "2-5", "6-10", "11-15", ">15"))) %>% 
  st_as_sf(., coords = c("Longitude", "Latitude"), crs = 4326)

# 2. Save the data ----

save(data_benthos_range, 
     file = "./../data/00_filtered/00-data_benthos-time_range_monitoring.RData")

```

# 5. Bind background maps in one file

```{r}

# 1. Load the maps ----

# 1.1 Major islands --

land_major <- read_sf("./../data/01_maps/ne_10m_land/ne_10m_land.shp", 
                      "ne_10m_land")

# 1.2 Minor islands --

land_minor <- read_sf("./../data/01_maps/ne_10m_minor_islands/ne_10m_minor_islands.shp", 
                      "ne_10m_minor_islands")
# 1.3 Reefs data --

land_reefs <- read_sf("./../data/01_maps/ne_10m_reefs/ne_10m_reefs.shp", "ne_10m_reefs")

# 2. Bind maps and save object ----

background_map <- bind_rows(land_reefs, land_minor) %>% 
  bind_rows(., land_major) %>% 
  st_as_sf(., crs = 4326) 

# 3. Save the data ----

save(background_map, file = "./../data/01_maps/00_background_map.RData")

```

# Reproducibility

```{r reprod}

# 1. Reproducibility ----

sessionInfo()

```

---
Nina PRASIL & Jeremy WICQUART | `r format(Sys.time())`
