==> R/01_analyses-maps.Rmd <==
---
title: "Analyses - Maps"
author : "Nina Prasil & Jeremy Wicquart"
date: "`r Sys.Date()`"
output: 
  html_document:
    theme: "cosmo"
    highlight: tango
    toc: true
    toc_float:
      collapsed: false
      smooth_scroll: false
    toc_depth: 4
---

```{r setup, include=FALSE}

knitr::opts_chunk$set(echo = TRUE, fig.align = "center")

```

# 1. Import functions, packages and data

```{r}

# 1. Source functions ----

source("functions/graphical_par.R")
source("functions/theme_map.R")
source("functions/ggplot_map_eez.R")
source("functions/ggplot_map_global.R")

# 2. Required packages ----

library(magrittr) # To use special pipes
library(tidyverse) # Core tidyverse packages
library(readxl) # To read excel files
library(sf) # To plot maps
sf_use_s2(FALSE) # Switch from S2 to GEOS
library(formattable) # For interactive html tables
library(DT) # For interactive html tables
library(ggpubr) # To assemble plots
library(ggsflabel)

# 3. Set default ggplot theme ----

theme_set(theme_map())

# 4. Load data ----

# 4.1 Data benthos - Raw (top 30) --

load("./../data/00_filtered/00-data_benthos-top_30.RData")

# 4.2 Data benthos - Summarized with time-range (top 30) --

load("./../data/00_filtered/00-data_benthos-time_range_monitoring.RData")

# 4.3 bbox file --

data_bbox <- read_xlsx("./../data/top_30_bbox.xlsx", sheet = 1)

# 4.4 EEZ data --

data_eez <- read_sf("../data/01_maps/World_EEZ_v11_20191118/eez_v11.shp") %>% 
  # Re-assign Territory depending to geopolitical issues
  mutate(SOVEREIGN1 = case_when(TERRITORY1 == "Glorioso Islands" ~ "France",
                                TERRITORY1 == "Chagos Archipelago" ~ "United Kingdom",
                                TRUE ~ SOVEREIGN1)) %>% 
  # Extract top 30 countries
  filter(SOVEREIGN1 %in% data_bbox$SOVEREIGN1) %>% 
  # Remove territories that do not contains coral reefs 
  filter(!(TERRITORY1 %in% c("Kerguélen", "Falkland / Malvinas Islands", "United Kingdom", "France",
                             "Saint Helena", "Alaska", "Crozet Islands", "Amsterdam and Saint Paul Islands",
                             "South Georgia and the South Sandwich Islands", "Jersey", "Guernsey", "French Guiana",
                             "Heard and McDonald Islands", "Macquarie Island", "Ascension",
                             "Saint-Pierre and Miquelon", "Tristan da Cunha", "Gibraltar")))

# 4.5 Background maps --

load(file = "./../data/01_maps/00_background_map.RData")

# 4.6 Background maps country boundaries --

land_boundary <- read_sf("../data/01_maps/ne_10m_admin_0_countries/ne_10m_admin_0_countries.shp")

# 4.7 Coral distribution --

coral_distribution <- read_sf("../data/02_reefs-at-risk_reef-data/reef_500_poly.shp") %>% 
  st_transform(crs = 4326) %>% 
  st_wrap_dateline()

```

# 2. Global maps

```{r}

# 1. Transform data to desired projection ----

# 1.1 Define the CRS of global maps (Equal Earth) --

crs_global <- "+proj=eqearth"

# 1.2 Transform the projection of EEZ data --

data_eez_trans <- data_eez %>% 
  st_transform(., crs = crs_global)

# 1.3 Transform the projection for background map --

background_map <- background_map %>% 
  st_transform(., crs = crs_global)

# 1.4 Create the background --

lats <- c(90:-90, -90:90, 90)
longs <- c(rep(c(180, -180), each = 181), 180)

world_map_background <- list(cbind(longs, lats)) %>%
  st_polygon() %>%
  st_sfc(crs = "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs") %>% 
  st_sf() %>%
  st_transform(crs = crs_global)

# 2. Create the plot for each of the countries ----

# 2.1 Base plot --

base_ggplot <- ggplot() +
  geom_sf(data = world_map_background, fill = "#56B4E950", color = "grey30", size = 0.5/.pt) +
  geom_sf(data = background_map, fill = col_fill_map, col = col_color_map) +
  theme(strip.background = element_rect(fill = "transparent", color = NA),
        plot.background = element_rect(fill = "transparent", color = NA),
        panel.grid = element_blank(),
        panel.background = element_blank(),
        panel.border = element_blank())

# 2.2 Loop to create all the plots --

for (i in as.character(unique(data_eez_trans$SOVEREIGN1))) {
  
  data_eez_i <- data_eez_trans %>% 
    filter(SOVEREIGN1 == i)
  
  plot_i <- base_ggplot +
    geom_sf(data = data_eez_i, col = col_eez_g, fill = col_eez_g, alpha = 0.7)
  
 ggsave(plot_i, 
        filename = paste0("../figs/01_global-map-by-country_", str_replace_all(tolower(i), " ", "-"), ".png"),
        height = 4)
  
}

# 3. Remove data ----

rm(world_map_background, plot_i, data_eez_i, crs_global, data_eez_trans)

```

# 3. Individual maps

## 3.1 Classic maps

```{r}

# 1. Create list of country to plot ----

data_bbox_classic <- data_bbox %>% 
  filter(plot_type == "classic") %>% 
  distinct(SOVEREIGN1) %>% 
  pull()

# 2. Make the loop ----

for (i in data_bbox_classic) {
  
  data_eez_i <- data_eez %>% 
    filter(SOVEREIGN1 == i)
  
  data_benthos_i <- data_benthos_range %>% 
    filter(SOVEREIGN1 == i)
  
  data_boundary_i <- data_bbox %>% 
    filter(SOVEREIGN1 == i)
  
  plot_i <- ggplot() +
    # EEZ
    geom_sf(data = data_eez_i, fill = col_eez, color = col_eez_b, alpha = 0.8) +
    # Coral reef distribution
    geom_sf(data = coral_distribution, col = "azure3", size = rel(0.2), alpha = 0.5, show.legend = FALSE) +
    # Background maps
    geom_sf(data = background_map, fill = col_fill_map, col = col_color_map) +
    geom_sf(data = land_boundary, fill = col_fill_map, col = col_color_map) +
    # Location sites
    geom_sf(data = data_benthos_i, shape = 21, col = "black", aes(fill = interval_class), size = 3) +
    # Set boundaries
    coord_sf(xlim = c(data_boundary_i$lon_min, data_boundary_i$lon_max), 
             ylim = c(data_boundary_i$lat_min, data_boundary_i$lat_max), expand = FALSE) +
    # Aesthetics
    guides(fill = guide_legend(override.aes = list(size = 4))) +
    scale_fill_manual(values = palette_trange, name = "Time range (years)", drop = FALSE) +
    theme(axis.title = element_blank())
  
  ggsave(plot_i, filename = paste0("../figs/02_individual-map-by-country_", str_replace_all(tolower(i), " ", "-"), ".png"))

}

# 3. Remove data ----

rm(data_bbox_classic, data_eez_i, data_benthos_i, data_boundary_i, plot_i)

```

## 3.2 Overlap maps

### 3.2.1 Fiji

```{r}

# 1. Filter and crop data ----

# 1.1 EEZ --

data_eez_i <- data_eez %>% 
  filter(SOVEREIGN1 == "Fiji") %>% 
  st_transform(., crs = 3460) %>% 
  st_union(.)

data_eez_i %<>% # Special pipe from magrittr
  st_buffer(10) # To join polygon (remove vertical line)

# 1.2 Background map --

background_map_i <- background_map %>% 
  st_transform(., crs = 3460) %>% 
  st_crop(st_bbox(data_eez_i)) %>% 
  filter(scalerank != 100) # To remove vertical line
  
# 1.3 Data benthos --

data_benthos_range_i <- data_benthos_range %>% 
  filter(SOVEREIGN1 == "Fiji") %>% 
  st_transform(., crs = 3460) %>% 
  st_crop(st_bbox(data_eez_i))

# 2. Make the plot ----

plot_i <- ggplot() +
  geom_sf(data = data_eez_i, fill = col_eez, color = col_eez_b, alpha = 0.8) +
  geom_sf(data = background_map_i, fill = col_fill_map, col = col_color_map) +
  geom_sf(data = data_benthos_range_i, shape = 21, col = "black", aes(fill = interval_class), size = 3) +
  # Aesthetics
  guides(fill = guide_legend(override.aes = list(size = 4))) +
  scale_fill_manual(values = palette_trange, name = "Time range (years)", drop = FALSE)

# 3. Save the plot ----

ggsave(plot_i, filename = "../figs/02_individual-map-by-country_fiji.png")

```

### 3.2.2 Tonga

```{r}

# 1. Filter and crop data ----

# 1.1 EEZ --

data_eez_i <- data_eez %>% 
  filter(SOVEREIGN1 == "Tonga") %>% 
  st_transform(., crs = 3460) %>% 
  st_union(.)

# 1.2 Background map --

background_map_i <- background_map %>% 
  st_transform(., crs = 3460) %>% 
  st_crop(st_bbox(data_eez_i))
  
# 1.3 Data benthos --

data_benthos_range_i <- data_benthos_range %>% 
  filter(SOVEREIGN1 == "Tonga") %>% 
  st_transform(., crs = 3460) %>% 
  st_crop(st_bbox(data_eez_i))

# 2. Make the plot ----

plot_i <- ggplot() +
  geom_sf(data = data_eez_i, fill = col_eez, color = col_eez_b, alpha = 0.8) +
  geom_sf(data = background_map_i, fill = col_fill_map, col = col_color_map) +
  geom_sf(data = data_benthos_range_i, shape = 21, col = "black", aes(fill = interval_class), size = 3) +
  # Aesthetics
  guides(fill = guide_legend(override.aes = list(size = 4))) +
  scale_fill_manual(values = palette_trange, name = "Time range (years)", drop = FALSE)

# 3. Save the plot ----

ggsave(plot_i, filename = "../figs/02_individual-map-by-country_tonga.png")

```

### 3.2.3 Kiribati

```{r}

# 1. Filter and crop data ----

# 1.1 EEZ --

data_eez_i <- data_eez %>% 
  filter(SOVEREIGN1 == "Kiribati") %>% 
  st_transform(., crs = "+proj=eqc +lat_ts=0 +lat_0=0 +lon_0=150 +x_0=0 +y_0=0 +datum=WGS84 +units=m +no_defs") %>% 
  st_buffer(0.05) %>% 
  st_union(.)

# 1.2 Background map --

background_map_i <- background_map %>% 
  st_transform(., crs = "+proj=eqc +lat_ts=0 +lat_0=0 +lon_0=150 +x_0=0 +y_0=0 +datum=WGS84 +units=m +no_defs") %>% 
  st_crop(st_bbox(data_eez_i))
  
# 1.3 Data benthos --

data_benthos_range_i <- data_benthos_range %>% 
  filter(SOVEREIGN1 == "Kiribati") %>% 
  st_transform(., crs = "+proj=eqc +lat_ts=0 +lat_0=0 +lon_0=150 +x_0=0 +y_0=0 +datum=WGS84 +units=m +no_defs") %>% 
  st_crop(st_bbox(data_eez_i))

# 2. Make the plot ----

plot_i <- ggplot() +
  geom_sf(data = data_eez_i, fill = col_eez, color = col_eez_b, alpha = 0.8) +
  geom_sf(data = background_map_i, fill = col_fill_map, col = col_color_map) +
  geom_sf(data = data_benthos_range_i, shape = 21, col = "black", aes(fill = interval_class), size = 3) +
  # Aesthetics
  guides(fill = guide_legend(override.aes = list(size = 4))) +
  scale_fill_manual(values = palette_trange, name = "Time range (years)", drop = FALSE)

# 3. Save the plot ----

ggsave(plot_i, filename = "../figs/02_individual-map-by-country_kiribati.png", height = 5)

```

## 3.3 Composed maps

### 3.3.1 France

```{r}

# 1. Filter territories -----

data_eez_sovereign <- data_eez %>% 
  filter(TERRITORY1 %in% c("French Polynesia", "Mayotte", "Ile Tromelin", "New Caledonia", 
                           "Réunion", "Europa Island", "Juan de Nova Island", "Bassas da India", 
                           "Clipperton Island", "Guadeloupe", "Saint-Barthélemy", "Glorioso Islands",
                           "Collectivity of Saint Martin", "Martinique", "Wallis and Futuna")) %>% 
  mutate(TERRITORY1 = str_replace_all(TERRITORY1, c("Ile Tromelin" = "Tromelin",
                                                    "Clipperton Island" = "Clipperton",
                                                    "Juan de Nova Island" = "Juan de Nova")))

# 2. Filter benthos time-range ----

data_benthos_range_i <- data_benthos_range %>% 
  filter(SOVEREIGN1 == "France")

# 3. Map for specific cases ----

# 3.1 Wallis and Futuna --

data_eez_i <- data_eez_sovereign %>% 
  filter(TERRITORY1 == "Wallis and Futuna")

plot_wafu <- ggplot() +
  # EEZ
  geom_sf(data = data_eez_i, fill = col_eez, color = col_eez_b) +
  # Background map
  geom_sf(data = background_map, fill = col_fill_map, col = col_color_map) +
  # Benthos
  geom_sf(data = data_benthos_range_i, shape = 21, col = "black", aes(fill = interval_class), size = 3, show.legend = FALSE) +
  coord_sf(xlim = c(-180, -174), ylim = c(-16, -10)) +
  scale_fill_manual(values = palette_trange, name = "Time range (years)", drop = FALSE) +
  theme(panel.background = element_blank(),
        panel.border = element_blank(),
        axis.title = element_blank(),
        axis.text = element_blank(),
        axis.ticks = element_blank(),
        plot.title = element_text(hjust = 0.5, size = 10),
        plot.margin = unit(c(0.1,0.1,0.1,0.1), "cm")) +
  labs(title = "Wallis and Futuna")

# 3.2 Bassas da India AND Europa Island --

data_eez_i <- data_eez_sovereign %>% 
  filter(TERRITORY1 %in% c("Bassas da India", "Europa Island"))

plot_baeu <- ggplot() +
  # EEZ
  geom_sf(data = data_eez_i, fill = col_eez, color = col_eez_b) +
  # Background map
  geom_sf(data = background_map, fill = col_fill_map, col = col_color_map) +
  # Benthos
  geom_sf(data = data_benthos_range_i, shape = 21, col = "black", aes(fill = interval_class), size = 3, show.legend = FALSE) +
  coord_sf(xlim = c(st_bbox(data_eez_i)$xmin, st_bbox(data_eez_i)$xmax), 
           ylim = c(st_bbox(data_eez_i)$ymin, st_bbox(data_eez_i)$ymax)) +  
  scale_fill_manual(values = palette_trange, name = "Time range (years)", drop = FALSE) +
  theme(panel.background = element_blank(),
        panel.border = element_blank(),
        axis.title = element_blank(),
        axis.text = element_blank(),
        axis.ticks = element_blank(),
        plot.title = element_text(hjust = 0.5, size = 10),
        plot.margin = unit(c(0.1,0.1,0.1,0.1), "cm")) +
  labs(title = "Bassas da India and Europa")

# 3.3 Guadeloupe AND Martinique --

data_eez_i <- data_eez_sovereign %>% 
  filter(TERRITORY1 %in% c("Guadeloupe", "Martinique"))

plot_magu <- ggplot() +
  # EEZ
  geom_sf(data = data_eez_i, fill = col_eez, color = col_eez_b) +
  # Background map
  geom_sf(data = background_map, fill = col_fill_map, col = col_color_map) +
  # Benthos
  geom_sf(data = data_benthos_range_i, shape = 21, col = "black", aes(fill = interval_class), size = 3, show.legend = FALSE) +
  coord_sf(xlim = c(st_bbox(data_eez_i)$xmin, st_bbox(data_eez_i)$xmax), 
           ylim = c(st_bbox(data_eez_i)$ymin, st_bbox(data_eez_i)$ymax)) +
  scale_fill_manual(values = palette_trange, name = "Time range (years)", drop = FALSE) +
  theme(panel.background = element_blank(),
        panel.border = element_blank(),
        axis.title = element_blank(),
        axis.text = element_blank(),
        axis.ticks = element_blank(),
        plot.title = element_text(hjust = 0.5, size = 10),
        plot.margin = unit(c(0.1,0.1,0.1,0.1), "cm")) +
  labs(title = "Guadeloupe and Martinique")

# 4. Make and assemble plots ----

plot_fr <- ggarrange(ggarrange(ggplot_map_eez(territory_name = "Collectivity of Saint Martin", title = TRUE),
                               ggplot_map_eez(territory_name = "Saint-Barthélemy", title = TRUE), 
                               plot_magu,
                               ggplot_map_eez(territory_name = "Glorioso Islands", title = TRUE), ncol = 4),
                     ggplot_map_global(), 
                     ggarrange(plot_wafu,
                               ggplot_map_eez(territory_name = "French Polynesia", title = TRUE),
                               ggplot_map_eez(territory_name = "Clipperton", title = TRUE),
                               ggplot_map_eez(territory_name = "New Caledonia", title = TRUE), ncol = 4), 
                     ggarrange(ggplot_map_eez(territory_name = "Mayotte", title = TRUE),
                               ggplot_map_eez(territory_name = "Juan de Nova", title = TRUE),
                               ggplot_map_eez(territory_name = "Réunion", title = TRUE),
                               plot_baeu,
                               ggplot_map_eez(territory_name = "Tromelin", title = TRUE), ncol = 5),
                     nrow = 4, heights = c(1, 2, 1, 1))

# 5. Save the plot ----

ggsave(plot_fr, file = "../figs/02_individual-map-by-country_france.png", width = 10, height = 12)

```

### 3.3.2 United States

```{r fig.width=9, fig.height=12}

# 1. Filter territories -----

data_eez_sovereign <- data_eez %>% 
  filter(TERRITORY1 %in% c("American Samoa", "Palmyra Atoll", "Puerto Rico", "United States Virgin Islands",
                           "Johnston Atoll", "Guam", "Northern Mariana Islands", "Wake Island", "Jarvis Island",
                           "Howland and Baker islands", "United States", "Hawaii")) %>% 
  filter(POL_TYPE == "200NM") %>% 
  mutate(TERRITORY1 = case_when(ISO_TER1 == "USA" ~ "South Coast of US",
                                TRUE ~ TERRITORY1),
         TERRITORY1 = str_replace_all(TERRITORY1, c("Howland and Baker islands" = "Howland and Baker Islands")))

# 2. Filter benthos time-range ----

data_benthos_range_i <- data_benthos_range %>% 
  filter(SOVEREIGN1 == "United States")

# 3. Map for specific cases ----

# 3.1 Puerto Rico and USVI --

data_eez_i <- data_eez_sovereign %>% 
  filter(TERRITORY1 %in% c("Puerto Rico", "United States Virgin Islands"))

plot_usvi <- ggplot() +
  # EEZ
  geom_sf(data = data_eez_i, fill = col_eez, color = col_eez_b) +
  # Background map
  geom_sf(data = background_map, fill = col_fill_map, col = col_color_map) +
  # Benthos
  geom_sf(data = data_benthos_range_i, shape = 21, col = "black", aes(fill = interval_class), size = 3, show.legend = FALSE) +
  coord_sf(xlim = c(-68.5, -63.5), ylim = c(14.5, 22)) +
  scale_fill_manual(values = palette_trange, name = "Time range (years)", drop = FALSE) +
  theme(panel.background = element_blank(),
        panel.border = element_blank(),
        axis.title = element_blank(),
        axis.text = element_blank(),
        axis.ticks = element_blank(),
        plot.title = element_text(hjust = 0.5, size = 10),
        plot.margin = unit(c(0.1,0.1,0.1,0.1), "cm")) +
  labs(title = "Puerto Rico and USVI")

# 3.2 South US coast --

data_eez_i <- data_eez_sovereign %>% 
  filter(TERRITORY1 == "United States")

plot_sous <- ggplot() +
  # EEZ
  geom_sf(data = data_eez_i, fill = col_eez, color = col_eez_b) +
  # Background map
  geom_sf(data = background_map, fill = col_fill_map, col = col_color_map) +
  # Benthos
  geom_sf(data = data_benthos_range_i, shape = 21, col = "black", aes(fill = interval_class), size = 3, show.legend = FALSE) +
  coord_sf(xlim = c(-94, -79), ylim = c(23.5, 31)) +
  scale_fill_manual(values = palette_trange, name = "Time range (years)", drop = FALSE) +
  theme(panel.background = element_blank(),
        panel.border = element_blank(),
        axis.title = element_blank(),
        axis.text = element_blank(),
        axis.ticks = element_blank(),
        plot.title = element_text(hjust = 0.5, size = 10),
        plot.margin = unit(c(0.1,0.1,0.1,0.1), "cm")) +
  labs(title = "South Coast of US")

# 3.3 Hawaii --

data_eez_i <- data_eez_sovereign %>% 
  filter(TERRITORY1 == "Hawaii")

plot_hawa <- ggplot() +
  # EEZ
  geom_sf(data = data_eez_i, fill = col_eez, color = col_eez_b) +
  # Background map
  geom_sf(data = background_map, fill = col_fill_map, col = col_color_map) +
  # Benthos
  geom_sf(data = data_benthos_range_i, shape = 21, col = "black", aes(fill = interval_class), size = 3, show.legend = FALSE) +
  coord_sf(xlim = c(-180, -150), ylim = c(16.5, 35)) +
  scale_fill_manual(values = palette_trange, name = "Time range (years)", drop = FALSE) +
  theme(panel.background = element_blank(),
        panel.border = element_blank(),
        axis.title = element_blank(),
        axis.text = element_blank(),
        axis.ticks = element_blank(),
        plot.title = element_text(hjust = 0.5, size = 10),
        plot.margin = unit(c(0.1,0.1,0.1,0.1), "cm")) +
  labs(title = "Hawaii")

# 3.4 Guam and Northern Mariana Islands --

data_eez_i <- data_eez_sovereign %>% 
  filter(TERRITORY1 %in% c("Guam", "Northern Mariana Islands"))

plot_guam <- ggplot() +
  # EEZ
  geom_sf(data = data_eez_i, fill = col_eez, color = col_eez_b) +
  # Background map
  geom_sf(data = background_map, fill = col_fill_map, col = col_color_map) +
  # Benthos
  geom_sf(data = data_benthos_range_i, shape = 21, col = "black", aes(fill = interval_class), size = 3, show.legend = FALSE) +
  coord_sf(xlim = c(141, 150), ylim = c(10, 25)) +
  scale_fill_manual(values = palette_trange, name = "Time range (years)", drop = FALSE) +
  theme(panel.background = element_blank(),
        panel.border = element_blank(),
        axis.title = element_blank(),
        axis.text = element_blank(),
        axis.ticks = element_blank(),
        plot.title = element_text(hjust = 0.5, size = 10),
        plot.margin = unit(c(0.1,0.1,0.1,0.1), "cm")) +
  labs(title = "Guam and Mariana Islands")

# 4. Make and assemble plots ----

plot_us <- ggarrange(ggarrange(plot_hawa,
                               ggplot_map_eez(territory_name = "Palmyra Atoll", title = TRUE),
                               plot_sous,
                               plot_usvi,
                               ncol = 4),
                     ggplot_map_global(), 
                     ggarrange(ggplot_map_eez(territory_name = "Johnston Atoll", title = TRUE),
                               ggplot_map_eez(territory_name = "American Samoa", title = TRUE),
                               ggplot_map_eez(territory_name = "Jarvis Island", title = TRUE), ncol = 3),
                     ggarrange(ggplot_map_eez(territory_name = "Howland and Baker Islands", title = TRUE),
                               plot_guam,
                               ggplot_map_eez(territory_name = "Wake Island", title = TRUE), ncol = 3),
                     nrow = 4, heights = c(1, 2, 1, 1))

# 5. Save the plot ----

ggsave(plot_us, file = "../figs/02_individual-map-by-country_united-states.png", width = 10, height = 12)

```

### 3.3.3 United Kingdom

```{r fig.width=9, fig.height=1}

# 1. Filter territories -----

data_eez_sovereign <- data_eez %>% 
  filter(TERRITORY1 %in% c("Pitcairn", "Anguilla", "British Virgin Islands",
                           "Cayman Islands", "Montserrat", "Turks and Caicos Islands",
                           "Bermuda", "Chagos Archipelago")) %>% 
  mutate(TERRITORY1 = str_replace_all(TERRITORY1, c("Chagos Archipelago" = "BIOT")))

# 2. Filter benthos time-range ----

data_benthos_range_i <- data_benthos_range %>% 
  filter(SOVEREIGN1 == "United Kingdom")

# 3. Make and assemble plots ----

plot_uk <- ggarrange(ggarrange(ggplot_map_eez(territory_name = "Turks and Caicos Islands", title = TRUE),
                     ggplot_map_eez(territory_name = "Cayman Islands", title = TRUE), 
                     ggplot_map_eez(territory_name = "BIOT", title = TRUE), 
                     ggplot_map_eez(territory_name = "Bermuda", title = TRUE), ncol = 4),
          ggplot_map_global(), 
          ggarrange(ggplot_map_eez(territory_name = "Pitcairn", title = TRUE),
                    ggplot_map_eez(territory_name = "Montserrat", title = TRUE),
                    ggplot_map_eez(territory_name = "Anguilla", title = TRUE),
                    ggplot_map_eez(territory_name = "British Virgin Islands", title = TRUE), ncol = 4), 
          nrow = 3, heights = c(1, 2, 1))

# 4. Save the plot ----

ggsave(plot_uk, file = "../figs/02_individual-map-by-country_united-kingdom.png", width = 10, height = 8)

```

# 4. Summary indicators

## 4.1 Transform data

```{r}

# 1. Transform indicators ----

data_indicators <- read_xlsx("../data/05_synthesis-indicators.xlsx") %>% 
  select(SOVEREIGN1, live_coral_cover, other_benthic_gps) %>% 
  left_join(., data_eez) %>%
  mutate_at(c("live_coral_cover", "other_benthic_gps"), 
            ~str_replace_all(., c("green" = "#68C3A3",
                                  "orange" = "#FF9416",
                                  "grey" = "#bdc3c7"))) %>% 
  select(SOVEREIGN1, live_coral_cover, other_benthic_gps, geometry) %>% 
  st_as_sf()

# 2. Transform subindicators ----

data_subindicators <- read_xlsx("../data/05_synthesis-indicators.xlsx") %>% 
  select(-live_coral_cover, -other_benthic_gps) %>% 
  pivot_longer(2:ncol(.), names_to = "subindicator", values_to = "color") %>% 
  mutate(type = ifelse(str_detect(subindicator, "lcc") == TRUE, "LCC", "OBG"),
         subindicator = str_remove_all(subindicator, "lcc_|obg_"),
         subindicator = str_replace_all(subindicator, c("spatial" = "Spatial extent",
                                                        "temp" = "Temporal extent",
                                                        "taxo" = "Taxonomic resolution",
                                                        "meth" = "Monitoring methods")),
         subindicator = as.factor(subindicator),
         subindicator = factor(subindicator, levels = c("Spatial extent", 
                                                        "Temporal extent", "Monitoring methods", 
                                                        "Taxonomic resolution")),
         color = str_replace_all(color, c("green" = "#68C3A3",
                                          "orange" = "#FF9416",
                                          "grey" = "#bdc3c7"))) %>% 
  group_by(type, color, subindicator) %>% 
  count(name = "abs") %>% 
  ungroup() %>% 
  group_by(subindicator) %>% 
  mutate(total = sum(abs),
         rel = (100*abs)/total)

# 3. Calculate percentage and export the table ----

read_xlsx("../data/05_synthesis-indicators.xlsx") %>% 
  pivot_longer(2:ncol(.), names_to = "indicator", values_to = "color") %>% 
  group_by(indicator, color) %>% 
  count() %>% 
  ungroup() %>% 
  group_by(indicator) %>% 
  mutate(rel = (n/sum(n))*100,
         n = paste0(n, " (", round(rel, 2), "%)")) %>% 
  ungroup() %>% 
  select(-rel) %>% 
  pivot_wider(names_from = color, values_from = n) %>% 
  write.csv2(., file = "../figs/05_indicators-summary_percentage.csv", row.names = FALSE)

```

## 4.2 Live coral cover

```{r}

# 1. Map of indicators ----

map_indicator <- base_ggplot +
  geom_sf(data = data_indicators, aes(fill = live_coral_cover), size = 0.5) +
  scale_fill_identity() +
  labs(title = "Live Coral Cover") +
  theme(plot.title = element_text(hjust = 0.5))

# 2. Donuts of subindicators ----

plot_donuts <- data_subindicators %>% 
  filter(type == "LCC") %>% 
  ggplot(data = ., aes(x = 2, y = rel, fill = color))+
  geom_bar(stat = "identity", show.legend = FALSE) +
  coord_polar("y", start = 200) +
  theme_void() +
  scale_fill_identity() +
  xlim(0.2,2.5) +
  facet_grid(~subindicator) +
  theme(strip.text.x = element_text(vjust = 1, family = font_choose_graph, size = 8))

# 3. Make the plot ----

ggarrange(map_indicator, plot_donuts, ncol = 1, heights = c(1.2, 0.5), labels = c("A", "B"))

# 4. Save the plot ----

ggsave(file = "../figs/05_indicators-summary_live-coral-cover.png", dpi = 600)

```

## 4.3 Other benthic groups

```{r}

# 1. Map of indicators ----

map_indicator <- base_ggplot +
  geom_sf(data = data_indicators, aes(fill = other_benthic_gps)) +
  scale_fill_identity() +
  labs(title = "Fleshy Algae Cover and Cover of Key Benthic Groups") +
  theme(plot.title = element_text(hjust = 0.5))

# 2. Donuts of subindicators ----

plot_donuts <- data_subindicators %>% 
  filter(type == "OBG") %>% 
  ggplot(data = ., aes(x = 2, y = rel, fill = color))+
  geom_bar(stat = "identity", show.legend = FALSE) +
  coord_polar("y", start = 200) +
  theme_void() +
  scale_fill_identity() +
  xlim(0.2,2.5) +
  facet_grid(~subindicator) +
  theme(strip.text.x = element_text(vjust = 1, family = font_choose_graph, size = 8))

# 3. Make the plot ----

ggarrange(map_indicator, plot_donuts, ncol = 1, heights = c(1.2, 0.5), labels = c("A", "B"))

# 4. Save the plot ----

ggsave(file = "../figs/05_indicators-summary_other-benthic-groups.png", dpi = 600)

```

# Reproducibility

```{r reprod}

# 1. Reproducibility ----

sessionInfo()

```

---
Nina PRASIL & Jeremy WICQUART | `r format(Sys.time())`
