==> R/02_analyses-graphs.Rmd <==
---
title: "Analyses - Graphs"
author : "Nina Prasil & Jeremy Wicquart"
date: "`r Sys.Date()`"
output: 
  html_document:
    theme: "cosmo"
    highlight: tango
    toc: true
    toc_float:
      collapsed: false
      smooth_scroll: false
    toc_depth: 4
---

```{r setup, include=FALSE}

knitr::opts_chunk$set(echo = TRUE, fig.align = "center")

```

# Import functions, packages and data

```{r base}

# 1. Source functions ----

source("functions/graphical_par.R")
source("functions/theme_graph.R")

# 2. Required packages ----

library(tidyverse) # Core tidyverse packages
library(sf) # To plot maps
library(formattable) # For interactive html tables
library(DT) # For interactive html tables
library(ggpubr) # To assemble plots

# 3. Set default ggplot theme ----

theme_set(theme_graph())

# 4. Import and modify data ----

load("./../data/00_filtered/00-data_benthos-top_30.RData")

data_benthos <- data_benthos %>% 
  mutate(Category = as.character(Category),
         Group = as.character(Group)) %>% 
  mutate(Longitude = st_coordinates(.)[,"X"],
         Latitude = st_coordinates(.)[,"Y"]) %>% 
  st_drop_geometry(.)

```

# Table reef area by country

```{r}

# 1. Load data ----

data_reefs <- read_csv2("../figs/04_table-reef-surface-by-country.csv")

# 2. Make the table ----

data_reefs %>% 
  formattable(.) %>% 
  as.datatable(., rownames = FALSE, colnames = c("N°", "Country", "Area (km²)", "Relative (%)"))

```

# Number of sites

```{r}

# 1. Full GCRMN dataset (all countries not only top 30) ----

data_benthos_all <- read.csv2("../data/03-merge_all_all_all_benthos_NA.csv") %>% 
  # Remove some datasets (those leading to bias)
  filter(!(DatasetID %in% c("XLCA1", "XLCA2", "XLCA3", "XLCA4", "XLCA5")))

# 2. Number of sites per country ----

nb_sites <- data_benthos %>% 
  select(SOVEREIGN1, Latitude, Longitude) %>% 
  distinct() %>% 
  group_by(SOVEREIGN1) %>% 
  summarise(n_sites = n())

# 3. Number of observations per country ----

nb_obs <- data_benthos %>% 
  group_by(SOVEREIGN1) %>% 
  summarise(n_obs_abs = n()) %>% 
  ungroup() %>% 
  # Number of observations relative to the overall dataset, not only top 30 country dataset
  mutate(n_obs_rel = (n_obs_abs*100)/(nrow(data_benthos_all))) %>% 
  arrange(SOVEREIGN1)

# 4. Export the table ----

left_join(nb_sites, nb_obs) %>% 
  write_csv2(., "../figs/04_table-monitoring-effort-by-country.csv")

# 5. Remove useless objects ----

rm(data_benthos_all, nb_obs, nb_sites)

```

# Table monitoring descriptors

```{r}

# 1. Number of sites per country ----

nb_sites <- data_benthos %>% 
  select(SOVEREIGN1, Latitude, Longitude) %>% 
  distinct() %>% 
  group_by(SOVEREIGN1) %>% 
  summarise(sites_abs = n()) %>% 
  ungroup() %>% 
  mutate(sites_rel = (sites_abs/sum(sites_abs))*100) %>% 
  left_join(., data_reefs) %>% 
  select(-pos, -total_area_rel) %>% 
  mutate(sites_100km = (sites_abs*100)/total_area_abs)

# 2. Number of surveys per country ----

nb_surveys <- data_benthos %>% 
  select(SOVEREIGN1, Latitude, Longitude, Year) %>% 
  distinct() %>% 
  group_by(SOVEREIGN1) %>% 
  summarise(surveys_abs = n()) %>% 
  ungroup() %>% 
  mutate(surveys_rel = (surveys_abs/sum(surveys_abs))*100) %>% 
  left_join(., data_reefs) %>% 
  select(-pos, -total_area_rel) %>% 
  mutate(surveys_100km = (surveys_abs*100)/total_area_abs)

# 3. Number of observations per country ----

nb_obs <- data_benthos %>% 
  group_by(SOVEREIGN1) %>% 
  summarise(obs_abs = n()) %>% 
  ungroup() %>% 
  mutate(obs_rel = (obs_abs/sum(obs_abs))*100) %>% 
  left_join(., data_reefs) %>% 
  select(-pos, -total_area_rel) %>% 
  mutate(obs_100km = (obs_abs*100)/total_area_abs)

# 4. Most used method ----

data_method <- data_benthos %>% 
  select(Latitude, Longitude, Year, Method, SOVEREIGN1) %>% 
  distinct(.) %>% 
  mutate(Method = str_split_fixed(Method, ",", 6)[,1],
         Method = str_to_lower(Method),
         Method = str_trim(Method, side = "both"),
         Method = str_squish(Method),
         Method = case_when(Method %in% c("line intercept transect", 
                                          "line intersect transect",
                                          "lit") ~ "LIT",
                            Method %in% c("point intercept method",
                                          "point intercept transect",
                                          "point intersect transect",
                                          "pit") ~ "PIT",
                            Method %in% c("photo-quadrat",
                                          "Photo-quadrat",
                                          "underwater photo transect") ~ "Photo-quadrat",
                            Method %in% c("visual census",
                                          "visual quadrat estimation",
                                          "vis est.") ~ "Visual census",
                            Method == "video-transect" ~ "Video transect",
                            TRUE ~ "Unknown")) %>% 
  group_by(Method, SOVEREIGN1) %>% 
  count(name = "Abs") %>% 
  ungroup() %>% 
  group_by(SOVEREIGN1) %>% 
  filter(Abs == max(Abs)) %>% 
  select(-Abs) %>% 
  rename(main_method = Method)

# 5. Percentage of observations at the Group level or above ----

precision_taxo <- data_benthos %>% 
   select(SOVEREIGN1, Year, Category, Group, Family, Genus, Species) %>% 
   mutate(Tax_lvl = case_when(!(is.na(Species)) ~ "Species",
                              !(is.na(Genus)) ~ "Genus",
                              !(is.na(Family)) ~ "Family",
                              !(is.na(Group)) ~ "Group",
                              !(is.na(Category)) ~ "Category")) %>% 
  mutate(Tax_lvl = str_replace_all(Tax_lvl, c("Group" = "Group_above",
                                              "Family" = "Group_above",
                                              "Genus" = "Group_above",
                                              "Species" = "Group_above"))) %>% 
  group_by(SOVEREIGN1, Tax_lvl) %>% 
  summarise(cat_abs = n()) %>% 
  ungroup() %>% 
  pivot_wider(names_from = Tax_lvl, values_from = cat_abs) %>% 
  mutate(percent_group = (Group_above*100)/(Category + Group_above)) %>% 
  select(-Category, -Group_above)

# 6. Number of years with data ----

nb_years <- data_benthos %>% 
  select(SOVEREIGN1, Year) %>% 
  distinct() %>% 
  group_by(SOVEREIGN1) %>% 
  summarise(n_years = n())

# 7. Combine and export the data ----

left_join(nb_sites, nb_surveys) %>% 
  left_join(., nb_obs) %>% 
  left_join(., data_method) %>% 
  left_join(., precision_taxo) %>%
  left_join(., nb_years) %>% 
  select(-total_area_abs) %>% 
  mutate_if(is.numeric, ~round(., digits = 2)) %>% 
  write_csv2(., "../figs/05_indicators-summary_monitoring-descriptors-by-country.csv")

# 8. Remove useless objects ----

rm(nb_sites, nb_surveys, nb_obs, data_method, precision_taxo, nb_years)

```

# Distribution of years

```{r}

# 1. Modify the data ----

data_years <- data_benthos %>% 
  select(Latitude, Longitude, Year, SOVEREIGN1) %>% 
  distinct(.) %>% 
  group_by(Year, SOVEREIGN1) %>% 
  count() %>% 
  group_by(SOVEREIGN1) %>% 
  mutate(Total = sum(n),
         freq = (n/sum(n))*100)

# 2. Make the loop to produce each plot ----

# 2.1 Create the empty list --

plot_years <- list() # Empty list

# 2.2 Loop --

for (i in unique(data_years$SOVEREIGN1)) {
  
  plot_i <- data_years %>% 
    filter(SOVEREIGN1 == i) %>% 
    ggplot(., aes(x = Year, y = freq)) +
      geom_col(fill = "#6c7a89", width = 0.8) +
      lims(x = c(1970, 2020)) +
      labs(x = "Year", y = "Percentage of surveys") +
      theme(plot.margin = margin(t = 0.75, b = 0.75, r = 0.75, l = 0.75, unit = "cm"))
  
  plot_years[[i]] <- plot_i # add each plot into plot list
  
}

```

# Methods used general

```{r}

# 1. Modify the data ----

data_method <- data_benthos %>% 
  select(Latitude, Longitude, Year, Method, SOVEREIGN1) %>% 
  distinct(.) %>% 
  mutate(Method = str_split_fixed(Method, ",", 6)[,1],
         Method = str_to_lower(Method),
         Method = str_trim(Method, side = "both"),
         Method = str_squish(Method),
         Method = case_when(Method %in% c("line intercept transect", 
                                          "line intersect transect",
                                          "lit") ~ "LIT",
                            Method %in% c("point intercept method",
                                          "point intercept transect",
                                          "point intersect transect",
                                          "pit") ~ "PIT",
                            Method %in% c("photo-quadrat",
                                          "Photo-quadrat",
                                          "underwater photo transect") ~ "Photo-quadrat",
                            Method %in% c("visual census",
                                          "visual quadrat estimation",
                                          "vis est.") ~ "Visual census",
                            Method == "video-transect" ~ "Video transect",
                            TRUE ~ "Unknown")) %>% 
  group_by(Method, SOVEREIGN1) %>% 
  count(name = "Abs") %>% 
  ungroup() %>% 
  complete(Method, SOVEREIGN1, fill = list(Abs = 0)) %>% 
  group_by(SOVEREIGN1) %>% 
  mutate(Rel = (Abs/sum(Abs))*100,
        Method = as.factor(Method),
        Method = factor(Method, 
                        levels = rev(c("Photo-quadrat", "LIT", "PIT", "Video transect", "Visual census", "Unknown")))) %>% 
  ungroup() %>% 
  arrange(SOVEREIGN1)

# 2. Make the loop to produce each plot ----

# 2.1 Create the empty list --

plot_method_general <- list() # Empty list

# 2.2 Loop --

for (i in unique(data_method$SOVEREIGN1)) {
  
  plot_i <- data_method %>% 
    filter(SOVEREIGN1 == i) %>% 
    ggplot(data = ., aes(x = Method, y = Rel, label = paste0(round(Rel, 0), " %"), fill = Method)) +
    geom_bar(stat = "identity", show.legend = FALSE) +
    geom_text(hjust = -0.2, family = font_choose_graph) +
    coord_flip() +
    theme(axis.text = element_text(size = 12),
          axis.title = element_text(size = 16),
          plot.tag = element_text(size = 20),
          plot.margin = margin(t = 0.75, b = 0.75, r = 0.75, l = 0.75, unit = "cm")) +
    scale_fill_manual(breaks = c("Photo-quadrat", "LIT", "PIT", "Video transect", "Visual census", "Unknown"),
                      values = c("#16a085", "#00b16a", "#00b5cc", "#3498db", "#19b5fe", "#6c7a89"),
                      drop = FALSE) +
    scale_y_continuous(name = "Percentage of surveys", 
                       breaks = c(0, 25, 50, 75, 100), 
                       limits = c(0,110)) +
    labs(x = NULL)
  
  plot_method_general[[i]] <- plot_i # add each plot into plot list
  
}

```

# Methods used by time-range

```{r}

# 1. Modify the data ----

data_method <- data_benthos %>% 
  select(Latitude, Longitude, Year, Method, SOVEREIGN1) %>% 
  distinct(.) %>% 
  mutate(Method = str_split_fixed(Method, ",", 6)[,1],
         Method = str_to_lower(Method),
         Method = str_trim(Method, side = "both"),
         Method = str_squish(Method),
         Method = case_when(Method %in% c("line intercept transect", 
                                          "line intersect transect",
                                          "lit") ~ "LIT",
                            Method %in% c("point intercept method",
                                          "point intercept transect",
                                          "point intersect transect",
                                          "pit") ~ "PIT",
                            Method %in% c("photo-quadrat",
                                          "Photo-quadrat",
                                          "underwater photo transect") ~ "Photo-quadrat",
                            Method %in% c("visual census",
                                          "visual quadrat estimation",
                                          "vis est.") ~ "Visual census",
                            Method == "video-transect" ~ "Video transect",
                            TRUE ~ "Unknown"),
         Time_range = case_when(Year >= 1970 & Year < 2000 ~ "1970-1999",
                                Year >= 2000 & Year < 2010 ~ "2000-2009",
                                Year >= 2010 & Year < 2020 ~ "2010-2019")) %>% 
  group_by(SOVEREIGN1, Method, Time_range) %>% 
  count(name = "Abs") %>% 
  drop_na(Time_range) %>% 
  ungroup() %>% 
  complete(SOVEREIGN1, Method, Time_range, fill = list(Abs = 0)) %>% 
  group_by(SOVEREIGN1) %>% 
  mutate(Rel = (Abs/sum(Abs))*100,
         Method = as.factor(Method),
         Method = factor(Method, levels = c("Photo-quadrat", "LIT", "PIT", "Video transect", "Visual census", "Unknown"))) %>% 
  ungroup() %>% 
  arrange(SOVEREIGN1)

# Make the total for the text above the bars

data_method_total <- data_method %>% 
  group_by(SOVEREIGN1, Time_range) %>% 
  summarise(Rel = sum(Rel),
            Abs = sum(Abs),
            Abs = paste0("n = ", Abs))

# 2. Make the loop to produce each plot ----

# 2.1 Create the empty list --

plot_method_range <- list() # Empty list

# 2.2 Loop --

for (i in unique(data_method$SOVEREIGN1)) {
  
  data_method_i <- data_method %>% 
    filter(SOVEREIGN1 == i)
    
  data_method_total_i <- data_method_total %>% 
    filter(SOVEREIGN1 == i)
  
  plot_i <- ggplot() +
    geom_bar(data = data_method_i, aes(x = Time_range, y = Rel, fill = Method), 
             stat = "identity", width = 0.7) +
    geom_text(data = data_method_total_i, aes(x = Time_range, y = Rel, label = Abs), vjust = -1, family = font_choose_graph) +
    scale_fill_manual(name = "Method",
                      breaks = c("Photo-quadrat", "LIT", "PIT", "Video transect", "Visual census", "Unknown"),
                      values = c("#16a085", "#00b16a", "#00b5cc", "#3498db", "#19b5fe", "#6c7a89"),
                      drop = FALSE) +
    scale_y_continuous(name = "Percentage of surveys", 
                       breaks = c(0, 25, 50, 75, 100), 
                       limits = c(0,110)) +
    labs(x = NULL, y = "Percentage of surveys") +
    theme(legend.position = "top",
          plot.margin = margin(t = 0.75, b = 0.75, r = 0.75, l = 0.75, unit = "cm")) + 
    guides(fill = guide_legend(title.position = "top", title.hjust = 0.5, nrow = 1))
  
  plot_method_range[[i]] <- plot_i # add each plot into plot list
  
}

# 3. Export the table ----

data_method %>% 
  select(-Rel) %>% 
  pivot_wider(names_from = "Method", values_from = "Abs") %>% 
  write_csv2(., "../figs/04_table-method-used-by-time-range-and-country.csv")

```

# Taxonomic precision by time-range

```{r}

# 1. Modify the data ----

precision_taxo <- data_benthos %>% 
   select(SOVEREIGN1, Year, Category, Group, Family, Genus, Species) %>% 
   mutate(Tax_lvl = case_when(!(is.na(Species)) ~ "Species",
                              !(is.na(Genus)) ~ "Genus",
                              !(is.na(Family)) ~ "Family",
                              !(is.na(Group)) ~ "Group",
                              !(is.na(Category)) ~ "Category"),
          Time_range = case_when(Year >= 1970 & Year < 2000 ~ "1970-1999",
                                 Year >= 2000 & Year < 2010 ~ "2000-2009",
                                 Year >= 2010 & Year < 2020 ~ "2010-2019")) %>% 
  group_by(SOVEREIGN1, Tax_lvl, Time_range) %>% 
  count(name = "Abs") %>% 
  drop_na(Time_range) %>% 
  ungroup() %>% 
  complete(SOVEREIGN1, Tax_lvl, Time_range, fill = list(Abs = 0)) %>% 
  group_by(SOVEREIGN1, Time_range) %>% 
  mutate(Rel = (Abs/sum(Abs))*100,
         Time_range = as.factor(Time_range),
         Tax_lvl = str_replace_all(Tax_lvl, c("Group" = "Subgroup",
                                              "Category" = "Group")),
         Tax_lvl = factor(Tax_lvl, levels = c("Group", "Subgroup", "Family", "Genus", "Species"))) %>% 
  ungroup() %>% 
  arrange(SOVEREIGN1)

# 2. Make the loop to produce each plot ----

# 2.1 Create the empty list --

plot_precision_general <- list() # Empty list

# 2.2 Loop --

for (i in unique(precision_taxo$SOVEREIGN1)) {
  
  plot_i <- precision_taxo %>% 
    filter(SOVEREIGN1 == i) %>% 
    ggplot(data = .) +
    geom_bar(aes(x = Time_range, y = Rel, fill = Tax_lvl), 
             stat = "identity", width = 0.7) +
    scale_fill_manual(name = "Taxonomic level", 
                      breaks = c("Group", "Subgroup", "Family", "Genus", "Species"),
                      values = c("#ffec8b", "#f7ca18", "#eb9532", "#e74c3c", "#cf000f")) +
    labs(x = NULL, y = "Percentage of rows") +
    theme(legend.position = "top",
          plot.margin = margin(t = 0.75, b = 0.75, r = 0.75, l = 0.75, unit = "cm")) + 
    guides(fill = guide_legend(title.position = "top", title.hjust = 0.5))
  
  plot_precision_general[[i]] <- plot_i # add each plot into plot list
  
}

# 3. Add taxonomic precision across all time-ranges ----

precision_taxo <- precision_taxo %>% 
  select(-Rel) %>% 
  group_by(SOVEREIGN1, Tax_lvl) %>% 
  summarise(Abs = sum(Abs)) %>% 
  ungroup() %>% 
  group_by(SOVEREIGN1) %>% 
  mutate(Rel = (Abs/sum(Abs))*100) %>% 
  ungroup() %>% 
  mutate(Time_range = "All") %>% 
  bind_rows(precision_taxo, .)

# 4. Export the table ----

precision_taxo %>% 
  select(-Abs) %>% 
  pivot_wider(names_from = "Tax_lvl", values_from = "Rel") %>% 
  select(SOVEREIGN1, Time_range, Group, Subgroup, Family, Genus, Species) %>% 
  arrange(SOVEREIGN1, Time_range) %>% 
  write_csv2(., "../figs/04_table-relative-taxonomic-precision-global.csv")

```

# Taxonomic precision algae

```{r}

# 1. Modify the data ----

taxo_algae <- data_benthos %>% 
  filter(Category == "Algae") %>% 
  mutate(Group = ifelse(is.na(Group), "Unknown", Group)) %>% 
  group_by(SOVEREIGN1, Group) %>% 
  count(name = "n_row") %>% 
  ungroup() %>% 
  group_by(SOVEREIGN1) %>% 
  mutate(total = sum(n_row)) %>% 
  ungroup() %>% 
  mutate(row_rel = (n_row*100)/total,
         Group = factor(Group, levels = c("Unknown", "Turf algae", "Macroalgae", "Coralline algae")))

# 2. Make the loop to produce each plot ----

# 2.1 Create the empty list --

plot_precision_algae <- list() # Empty list

# 2.2 Loop --

for (i in unique(taxo_algae$SOVEREIGN1)) {
  
  plot_i <- taxo_algae %>% 
    filter(SOVEREIGN1 == i) %>% 
    ggplot(data = ., aes(x = SOVEREIGN1, y = row_rel, fill = Group)) +
      geom_bar(stat = "identity") +
      coord_flip() + 
      scale_fill_manual(name = "Subgroups of algae",
                        breaks = c("Coralline algae", "Macroalgae", "Turf algae", "Unknown"),
                        values = c("#e08283", "#00b16a", "#23cba7", "#6c7a89"),
                        drop = FALSE) +
      theme(axis.text.y = element_blank(),
            axis.ticks.y = element_blank(),
            axis.title.y = element_blank(),
            panel.grid = element_blank(),
            plot.margin = margin(t = 0.75, b = 0.75, r = 0.75, l = 0.75, unit = "cm"),
            panel.border = element_rect(colour = "white"),
            axis.line.x = element_line(colour = "black"),
            legend.position = "top") +
      scale_y_continuous(limits = c(0, 100), 
                         expand = c(0,0), 
                         labels = function(x) paste0(x, " %")) +
      labs(y = "Percentage of rows") + 
      guides(fill = guide_legend(title.position = "top", title.hjust = 0.5))
  
  plot_precision_algae[[i]] <- plot_i # add each plot into plot list
  
}

# 3. Export the table ----

taxo_algae %>% 
  select(-n_row, -total) %>% 
  pivot_wider(names_from = "Group", values_from = "row_rel") %>% 
  write_csv2(., "../figs/04_table-relative-taxonomic-precision-algae.csv")

```

# Taxonomic precision other fauna

```{r}

# 1. Modify the data ----

taxo_othfau <- data_benthos %>% 
  filter(Category == "Other fauna") %>% 
  mutate(Group = ifelse(Group %in% c("Porifera", "Tunicata", "Alcyonacea") | is.na(Group), Group, "Other subgroups"),
         Group = ifelse(is.na(Group), "Unknown", Group)) %>% 
  group_by(SOVEREIGN1, Group) %>% 
  count(name = "n_row") %>% 
  ungroup() %>% 
  group_by(SOVEREIGN1) %>% 
  mutate(total = sum(n_row)) %>% 
  ungroup() %>% 
  mutate(row_rel = (n_row*100)/total,
         Group = factor(Group, levels = c("Unknown", "Tunicata", "Porifera", "Other subgroups", "Alcyonacea")))

# 2. Make the loop to produce each plot ----

# 2.1 Create the empty list --

plot_precision_othfau <- list() # Empty list

# 2.2 Loop --

for (i in unique(taxo_othfau$SOVEREIGN1)) {
  
  plot_i <- taxo_othfau %>% 
    filter(SOVEREIGN1 == i) %>% 
    ggplot(data = ., aes(x = SOVEREIGN1, y = row_rel, fill = Group)) +
      geom_bar(stat = "identity") +
      coord_flip() + 
      scale_fill_manual(name = "Subgroups of other fauna",
                        breaks = c("Alcyonacea", "Other subgroups", "Porifera", "Tunicata", "Unknown"),
                        values = c("#e47833", "#f4b350", "#947cb0", "#fff9de", "#6c7a89"),
                        drop = FALSE) +
      theme(axis.text.y = element_blank(),
            axis.ticks.y = element_blank(),
            axis.title.y = element_blank(),
            panel.grid = element_blank(),
            plot.margin = margin(t = 0.75, b = 0.75, r = 0.75, l = 0.75, unit = "cm"),
            panel.border = element_rect(colour = "white"),
            axis.line.x = element_line(colour = "black"),
            legend.position = "top") +
      scale_y_continuous(limits = c(0, 100), 
                         expand = c(0,0), 
                         labels = function(x) paste0(x, " %")) +
      labs(y = "Percentage of rows") + 
      guides(fill = guide_legend(title.position = "top", title.hjust = 0.5))
  
  plot_precision_othfau[[i]] <- plot_i # add each plot into plot list
  
}

# 3. Export the table ----

taxo_othfau %>% 
  select(-n_row, -total) %>% 
  pivot_wider(names_from = "Group", values_from = "row_rel") %>% 
  write_csv2(., "../figs/04_table-relative-taxonomic-precision-other-fauna.csv")

```

# Assemblage of plots

```{r}

# 1. For loop to combine and export all the plots ----

for (i in names(plot_method_general)) {
  
  plot_i <- ggarrange(plotlist = c(plot_years[i], plot_method_general[i], 
                       plot_method_range[i], plot_precision_general[i],
                       plot_precision_algae[i], plot_precision_othfau[i]), 
          ncol = 2, nrow = 3, heights = c(0.9, 1.1, 0.5), 
          labels = c("A", "B", "C", "D", "E", "F"),
          font.label = list(size = 21, color = "black"))
  
  ggsave(plot = plot_i, 
         filename = paste0("../figs/03_graph-description-method-combined_", 
                str_replace_all(tolower(i), " ", "-"), ".png"),
         width = 14, height = 13)
  
}

# Issue on legends when using extrafont, see : 
# https://stackoverflow.com/questions/36425003/using-r-package-extrafont-is-generating-error-font-width-unknown-for-character

```

# Reproducibility

```{r reprod}

# 1. Reproducibility ----

sessionInfo()

```

---
Nina PRASIL & Jeremy WICQUART | `r format(Sys.time())`
