==> R/05_analyses-mpa.Rmd <==
---
title: "Analyses - MPA by country"
author : "Nina Prasil & Jeremy Wicquart"
date: "`r Sys.Date()`"
output: 
  html_document:
    theme: "cosmo"
    highlight: tango
    toc: true
    toc_float:
      collapsed: false
      smooth_scroll: false
    toc_depth: 4
---

```{r setup, include=FALSE}

knitr::opts_chunk$set(echo = TRUE, fig.align = "center")

```

# 1. Import functions, packages and data
- MPA zones implemented in the 30 countries studied
- Data filtration : 
  * Were retained only protected areas fitting the "is MAP" definition (as defined by MPAtlas) 
  * Are taken into account only MPA zones that are considered implemented (as defined by MPAtlas)
- Simplified geometries for MPA polygons (100m tolerance)

```{r base}

##### 1. Required packages ----

library(tidyverse) # Core tidyverse packages
#library(smoothr) # To smooth and tidy spatial features
library(sf) # To plot maps
sf_use_s2(FALSE) # Switch from S2 to GEOS


##### 2. Load data ----

### 2.1. MPAs from MPAtlas

data_mpa <- read_sf("./../data/02_mpa/mpatlas_20201223_clean/mpatlas_20201223_clean.shp") %>% 
  # Select top 30
  filter(sovereign %in% c("AUS", "IDN", "PHL", "FRA", "PNG", "USA", "FJI", "SLB", "SAU", "MDV",
                          "CUB", "FSM", "BHS", "MDG", "MHL", "IND", "MYS", "EGY", "KIR", "TZA",
                          "MOZ", "SDN", "ERI", "JPN", "SYC", "BLZ", "GBR", "VUT", "TON", "MEX")) %>% 
  # Homogenize countries names
  rename(SOVEREIGN1 = sovereign) %>% 
  filter(is_mpa == 1) %>% 
  filter(implemente == 1) %>% 
  mutate(SOVEREIGN1 = str_replace_all(SOVEREIGN1, 
                                      c("AUS" = "Australia",
                                        "BHS" = "Bahamas", 
                                        "BLZ" = "Belize",
                                        "CUB" = "Cuba", 
                                        "EGY" = "Egypt", 
                                        "ERI" = "Eritrea", 
                                        "FJI" = "Fiji", 
                                        "FRA" = "France", 
                                        "FSM" = "Micronesia",
                                        "GBR" = "United Kingdom",
                                        "IDN" = "Indonesia", 
                                        "IND" = "India", 
                                        "JPN" = "Japan", 
                                        "KIR" = "Kiribati",
                                        "MDG" = "Madagascar", 
                                        "MDV" = "Maldives", 
                                        "MEX" = "Mexico",
                                        "MHL" = "Marshall Islands",
                                        "MOZ" = "Mozambique", 
                                        "MYS" = "Malaysia",
                                        "PHL" = "Philippines", 
                                        "PNG" = "Papua New Guinea",
                                        "SAU" = "Saudi Arabia",
                                        "SDN" = "Sudan", 
                                        "SLB" = "Solomon Islands",
                                        "SYC" = "Seychelles",
                                        "TON" = "Tonga",
                                        "TZA" = "Tanzania",
                                        "USA" = "United States",
                                        "VUT" = "Vanuatu"))) %>% 
  st_transform(., crs = 32663) %>% 
  st_simplify(., dTolerance = 100) %>% 
  st_make_valid(.)



## 2.2 Selecting the pilot countries

top_30 <- unique(data_mpa$SOVEREIGN1)



## 2.3 EEZ from Flanders Institute 

data_eez <- read_sf("../data/01_maps/World_EEZ_v11_20191118/eez_v11.shp") %>% 
  # Re-assign Territory depending to geopolitical issues
  mutate(SOVEREIGN1 = case_when(TERRITORY1 == "Glorioso Islands" ~ "France",
                                TERRITORY1 == "Chagos Archipelago" ~ "United Kingdom",
                                TRUE ~ SOVEREIGN1)) %>% 
  # Extract top 30 countries
  filter(SOVEREIGN1 %in% top_30) %>% 
  # Remove territories that do not contains coral reefs 
  filter(!(TERRITORY1 %in% c("Kerguélen", "Falkland / Malvinas Islands", "United Kingdom", "France",
                             "Saint Helena", "Alaska", "Crozet Islands", "Amsterdam and Saint Paul Islands",
                             "South Georgia and the South Sandwich Islands", "Jersey", "Guernsey", "French Guiana",
                             "Heard and McDonald Islands", "Macquarie Island", "Ascension",
                             "Saint-Pierre and Miquelon", "Tristan da Cunha", "Gibraltar"))) %>% 
  st_transform(., crs = 32663) %>% 
  st_simplify(., dTolerance = 100) %>% 
  st_make_valid()




### 2.4 Transformed data from WRI - Reefs merged with EEZ 
# 500 m resolution

load(file = "./../data/reefs.eez.RData")


data_reefs <- reefs.eez %>%
  # Re-assign TERRITORY1 depending on geopolitical issues
  mutate(SOVEREIGN1 = case_when(TERRITORY1 == "Glorioso Islands" ~ "France",
                                TERRITORY1 == "Chagos Archipelago" ~ "United Kingdom",
                                TRUE ~ SOVEREIGN1)) %>% 
  st_transform(., crs = 32663)

remove(reefs.eez)

### 2.5 Surface of coral reef per country

reef_total <- read_csv2("../figs/04_table-reef-surface-by-country.csv")

```

# 2. Total surface of MPA per country

```{r}

## 1. Joining the MPA polygons and EEZ to exclude the terrestrial part of the MPAs
# CRS = 4326 ; simplified geometry ; merged with EEZ 

##For all countries except Australia (intersection pb with EEZ)
mpa_tot_km2 <- data_mpa %>%
  filter(SOVEREIGN1 != "Australia") %>% 
  st_intersection(., data_eez) %>% 
  group_by(SOVEREIGN1) %>% 
  summarize(geometry = st_union(st_combine(geometry), by_feature = TRUE)) %>%
  st_transform(., crs = 4326) %>% 
  mutate(mpa_area = st_area(geometry)) %>% 
  st_drop_geometry() %>% 
  ungroup()

units(mpa_tot_km2$mpa_area) <- "km2"

## For Australia

mpa_aus <- data_mpa %>% 
  filter(SOVEREIGN1 == "Australia") %>% 
  st_make_valid(.) %>% 
  st_intersection(., data_eez) %>% 
  st_make_valid() %>% 
  st_buffer(., dist = 0.1) %>% 
  summarize(geometry = st_union(geometry)) 

mpa_aus_km2 <- mpa_aus %>% 
  #st_make_valid(.) %>% 
  st_transform(., crs = 4326) %>% 
  #st_make_valid() %>% 
  mutate(mpa_area = st_area(geometry)) %>% 
  st_drop_geometry() 

units(mpa_aus_km2$mpa_area) <- "km2"

mpa_tot_km2 <- mpa_tot_km2 %>% 
  bind_rows(., mpa_aus_km2) %>% 
  mutate(SOVEREIGN1 = replace_na(SOVEREIGN1, "Australia")) %>% 
  arrange(SOVEREIGN1) 



# 3. Remove intermediate variables

remove(mpa_surface)

# 4. Export table including the surface of MPA per country

write_csv2(mpa_tot_km2, "../figs/00_mpa-with-reef.csv")


```

# 3. Reef surface under protection

```{r}

#### 1. Dissolve all MPA zones in one polygon per country

top_30 <- unique(data_mpa$SOVEREIGN1)

total_mpa <- data_mpa %>% 
  group_by(SOVEREIGN1) %>% 
  summarize(geometry = st_union(st_combine(geometry), by_feature = TRUE)) %>% 
  mutate(mpa_area = st_area(geometry)) %>% 
  ungroup()


# Export mpa_tot, since spatial analyses are time consuming
save(total_mpa, 
     file = "./../data/02_mpa/00-data_mpa_union.RData")

#total_mpa <- st_make_valid(total_mpa)

#### 2. Coral reefs intersection with MPA

coral_mpa <- data_reefs %>% 
  filter(SOVEREIGN1 %in% top_30) %>% 
  st_make_valid(.) %>% 
  st_intersection(., total_mpa) %>%   
  mutate(cr_mpa = st_area(geometry)) %>% 
  st_drop_geometry() %>% 
  group_by(SOVEREIGN1) %>% 
  summarise(coral_mpa_km2 = sum(cr_mpa/1000000)) %>% 
  arrange(SOVEREIGN1)  


# 3. Percentage of coral reefs under protection

protected_reefs_km2 <- coral_mpa %>% 
  mutate(coral_mpa_km2 = as.numeric(coral_mpa_km2)) %>%  
  left_join(reef_total, .) %>% 
  group_by(SOVEREIGN1) %>% 
  mutate(coral_mpa_rel = coral_mpa_km2/(total_area_abs)*100) %>% 
  rename(., coral_area_abs = total_area_abs, 
         coral_area_rel = total_area_rel) 



# Export the results: coral reef area 
write_csv2(protected_reefs_km2, "./../figs/06_protected-coral-reefs.csv")


```


# Reproducibility

```{r reprod}

# 1. Reproducibility ----

sessionInfo()

```

---
Nina PRASIL & Jeremy WICQUART | `r format(Sys.time())`
